Renaming error type file
